DjangoWeb/requirement.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.11.8
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
filelock==3.0.12
Pillow==8.0.1
pipenv==2020.11.4
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.1.0
virtualenv-clone==0.5.4
